# specs/api/workspace-sync.openapi.yaml
openapi: 3.1.0
info:
  title: MDR Workspace Sync API (Draft)
  version: 0.1.0
  description: >
    Draft API for workspace-first editing. The virtual document tree is internal
    and system-managed. Clients send user-level intents from Blueprint actions.
    Status: Draft-Frozen (API-003), Date: 2026-02-08.
servers:
  - url: http://localhost:8080
paths:
  /api/workspaces/{workspaceId}:
    get:
      summary: Get workspace snapshot
      operationId: getWorkspace
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Workspace snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetWorkspaceResponse'
  /api/workspaces/{workspaceId}/capabilities:
    get:
      summary: Get supported intent and command capabilities
      operationId: getWorkspaceCapabilities
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Capability map
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetCapabilitiesResponse'
  /api/workspaces/{workspaceId}/documents/{documentId}:
    patch:
      summary: Patch one document with a command
      description: >
        Apply one versioned CommandEnvelope to a MIR v1.3 graph-only document.
        Document content changes are PATCH-only: the server applies
        command.forwardOps, verifies command.reverseOps can restore the previous
        document, validates the patched MIR, advances the document content
        revision, and records the command in the operation log. Paths under
        /ui/root and the document root path / are forbidden.
      operationId: patchDocument
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
        - in: path
          name: documentId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchDocumentRequest'
      responses:
        '200':
          description: Patched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationSuccessResponse'
        '422':
          description: Invalid command, forbidden patch path, or MIR validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/workspaces/{workspaceId}/intents:
    post:
      summary: Apply one user intent
      operationId: applyWorkspaceIntent
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyIntentRequest'
      responses:
        '200':
          description: Applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationSuccessResponse'
        '422':
          description: Unsupported or invalid intent envelope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/workspaces/{workspaceId}/batch:
    post:
      summary: Apply a batch of document patches and intents
      operationId: applyWorkspaceBatch
      parameters:
        - in: path
          name: workspaceId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplyBatchRequest'
      responses:
        '200':
          description: Applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationSuccessResponse'
        '422':
          description: Unsupported operation or invalid envelope in batch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  schemas:
    GetWorkspaceResponse:
      type: object
      required: [workspace]
      properties:
        workspace:
          $ref: '#/components/schemas/WorkspaceSnapshot'
    GetCapabilitiesResponse:
      type: object
      required: [workspaceId, capabilities]
      properties:
        workspaceId:
          type: string
        capabilities:
          type: object
          description: >
            Key format namespace.type@version. Includes both intent and command
            capabilities. Reserved domains (e.g. core.nodegraph.*,
            core.animation.*) may appear as false before corresponding editors
            are implemented.
          additionalProperties:
            type: boolean
    WorkspaceSnapshot:
      type: object
      required:
        [id, workspaceRev, routeRev, opSeq, tree, documents, routeManifest]
      properties:
        id:
          type: string
        workspaceRev:
          type: integer
          minimum: 1
        routeRev:
          type: integer
          minimum: 1
        opSeq:
          type: integer
          minimum: 1
        tree:
          $ref: '#/components/schemas/WorkspaceTree'
        documents:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceDocument'
        routeManifest:
          type: object
          additionalProperties: true
    WorkspaceTree:
      type: object
      required: [rootId, nodes]
      properties:
        rootId:
          type: string
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/WorkspaceTreeNode'
    WorkspaceTreeNode:
      type: object
      required: [id, kind, name, parentId]
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [dir, doc]
        name:
          type: string
        parentId:
          type: [string, 'null']
        children:
          type: array
          items:
            type: string
        docId:
          type: string
    WorkspaceDocument:
      type: object
      required: [id, type, path, contentRev, metaRev, content]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [mir-page, mir-layout, mir-component, mir-graph, mir-animation]
        path:
          type: string
        contentRev:
          type: integer
          minimum: 1
        metaRev:
          type: integer
          minimum: 1
        content:
          type: object
          additionalProperties: true
        updatedAt:
          type: string
          format: date-time
    PatchDocumentRequest:
      type: object
      required: [expectedContentRev, command]
      properties:
        expectedContentRev:
          type: integer
          minimum: 1
        command:
          $ref: '#/components/schemas/CommandEnvelope'
        clientMutationId:
          type: string
    ApplyIntentRequest:
      type: object
      required: [expectedWorkspaceRev, intent]
      properties:
        expectedWorkspaceRev:
          type: integer
          minimum: 1
        expectedRouteRev:
          type: integer
          minimum: 1
        intent:
          $ref: '#/components/schemas/IntentEnvelope'
        clientMutationId:
          type: string
    IntentEnvelope:
      type: object
      required: [id, namespace, type, version, payload, issuedAt]
      properties:
        id:
          type: string
        namespace:
          type: string
          description: core | plugin.<vendor> | custom.<vendor> | core.nodegraph | core.animation
        type:
          type: string
        version:
          type: string
          description: semantic version string, e.g. 1.0
        payload:
          type: object
          additionalProperties: true
        idempotencyKey:
          type: string
        actor:
          type: object
          properties:
            userId:
              type: string
            clientId:
              type: string
          additionalProperties: false
        issuedAt:
          type: string
          format: date-time
    PatchOp:
      type: object
      required: [op, path]
      description: >
        Restricted JSON Patch operation. Document PATCH allows paths under
        /ui/graph, /logic, /animation, /metadata, and /x-* only. /ui/root and /
        are forbidden. Array "-" is only valid for add.
      properties:
        op:
          type: string
          enum: [add, remove, replace, move, copy, test]
        path:
          type: string
        from:
          type: string
        value:
          description: Operation payload
      additionalProperties: false
    CommandEnvelope:
      type: object
      required:
        [id, namespace, type, version, issuedAt, forwardOps, reverseOps, target]
      properties:
        id:
          type: string
        namespace:
          type: string
        type:
          type: string
        version:
          type: string
          description: semantic version string, e.g. 1.0
        issuedAt:
          type: string
          format: date-time
        forwardOps:
          type: array
          items:
            $ref: '#/components/schemas/PatchOp'
        reverseOps:
          type: array
          items:
            $ref: '#/components/schemas/PatchOp'
        target:
          type: object
          required: [workspaceId]
          properties:
            workspaceId:
              type: string
            documentId:
              type: string
          additionalProperties: false
        mergeKey:
          type: string
    ApplyBatchRequest:
      type: object
      required: [expectedWorkspaceRev, operations]
      properties:
        expectedWorkspaceRev:
          type: integer
          minimum: 1
        expectedRouteRev:
          type: integer
          minimum: 1
        operations:
          type: array
          items:
            oneOf:
              - $ref: '#/components/schemas/BatchPatchDocumentOperation'
              - $ref: '#/components/schemas/BatchIntentOperation'
        clientBatchId:
          type: string
    BatchPatchDocumentOperation:
      type: object
      required: [op, documentId, expectedContentRev, command]
      properties:
        op:
          type: string
          const: patchDocument
        documentId:
          type: string
        expectedContentRev:
          type: integer
          minimum: 1
        command:
          $ref: '#/components/schemas/CommandEnvelope'
    BatchIntentOperation:
      type: object
      required: [op, intent]
      properties:
        op:
          type: string
          const: intent
        intent:
          $ref: '#/components/schemas/IntentEnvelope'
    MutationSuccessResponse:
      type: object
      required: [workspaceId, workspaceRev, routeRev, opSeq]
      properties:
        workspaceId:
          type: string
        workspaceRev:
          type: integer
        routeRev:
          type: integer
        updatedDocuments:
          type: array
          items:
            type: object
            required: [id, contentRev, metaRev]
            properties:
              id:
                type: string
              contentRev:
                type: integer
              metaRev:
                type: integer
            additionalProperties: false
        opSeq:
          type: integer
        acceptedMutationId:
          type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorPayload'
    ErrorPayload:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples:
            - WKS-4001
            - WKS-4002
            - WKS-4003
            - WKS-5001
            - WKS-5002
            - MIR-4001
            - API-1001
        severity:
          type: string
          enum: [info, warning, error, fatal]
        domain:
          type: string
        retryable:
          type: boolean
        requestId:
          type: string
        docsUrl:
          type: string
        message:
          type: string
        details:
          description: Safe machine-readable context. Revision conflict details include conflictType, workspaceId, serverWorkspaceRev, serverRouteRev, opSeq, and optionally serverDocument.
        diagnostics:
          type: array
          items:
            $ref: '#/components/schemas/BackendDiagnostic'
    BackendDiagnostic:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, error, fatal]
        domain:
          type: string
        retryable:
          type: boolean
        path:
          type: string
        targetRef:
          type: object
          additionalProperties: true
        docsUrl:
          type: string
        details:
          additionalProperties: true
      additionalProperties: true